CR Fix: separate module file
