+++ lib/rspec/rest/dsl.rb
@@ -2,6 +2,7 @@
 
 require_relative "config"
 require_relative "errors"
+require_relative "expectations"
 require_relative "session"
 module RSpec
   module Rest
@@ -93,8 +94,9 @@ module RSpec
         end
       end
 
-      # rubocop:disable Metrics/ModuleLength
       module InstanceMethods
+        include Expectations
+
         def rest_session
           @rest_session ||= Session.new(self.class.rest_config)
         end
@@ -109,63 +111,6 @@ module RSpec
           ensure_request_context!
           execute_rest_request_if_pending
           rest_session.last_request
-        end
-
-        def expect_status(code)
-          expect(rest_response.status).to eq(code)
-        end
-
-        def expect_header(key, value_or_regex)
-          actual = header_value_for(key)
-          available_keys = rest_response.headers.keys.map(&:to_s).sort.join(", ")
-          message = "Expected response header #{key.inspect} to be present. " \
-                    "Available headers: [#{available_keys}]"
-          raise ::RSpec::Expectations::ExpectationNotMetError, message if actual.nil?
-
-          if value_or_regex.is_a?(Regexp)
-            expect(actual).to match(value_or_regex)
-          else
-            expect(actual).to eq(value_or_regex)
-          end
-        end
-
-        def expect_json(expected = nil, &block)
-          parsed = rest_response.json
-
-          if block
-            instance_exec(parsed, &block)
-            return parsed
-          end
-
-          return parsed if expected.nil?
-
-          if expected.respond_to?(:matches?)
-            expect(parsed).to expected
-          else
-            expect(parsed).to eq(expected)
-          end
-
-          parsed
-        end
-
-        def integer
-          be_a(Integer)
-        end
-
-        def string
-          be_a(String)
-        end
-
-        def boolean
-          satisfy("be boolean") { |value| [true, false].include?(value) }
-        end
-
-        def array_of(matcher)
-          all(matcher)
-        end
-
-        def hash_including(*)
-          a_hash_including(*)
         end
 
         def header(key, value)
@@ -249,17 +194,6 @@ module RSpec
                 "No active REST request context. Call this inside a verb block (get/post/put/patch/delete)."
         end
 
-        def header_value_for(key)
-          headers = rest_response.headers
-          return headers[key] if headers.key?(key)
-
-          key_str = key.to_s
-          pair = headers.find do |header_key, _|
-            header_key.to_s.casecmp(key_str).zero?
-          end
-          pair&.last
-        end
-
         def apply_path_params(path, params)
           rendered = params.reduce(path.to_s) do |current, (key, value)|
             current.gsub("{#{key}}", value.to_s)
@@ -275,7 +209,6 @@ module RSpec
           rendered
         end
       end
-      # rubocop:enable Metrics/ModuleLength
 
       def self.included(base)
         base.extend(ClassMethods)
